chore: separate provided, java build tests into al2023
