+++ tests/integration/buildcmd/test_build_cmd.py
@@ -26,8 +26,6 @@ from tests.testing_utils import (
     SKIP_DOCKER_BUILD,
     SKIP_DOCKER_MESSAGE,
     UpdatableSARTemplate,
-    runtime_supported_by_docker,
-    RUNTIME_NOT_SUPPORTED_BY_DOCKER_MSG,
 )
 from tests.integration.buildcmd.build_integ_base import (
     BuildIntegBase,
@@ -36,12 +34,8 @@ from tests.integration.buildcmd.build_integ_base import (
     BuildIntegRubyBase,
     NestedBuildIntegBase,
     IntrinsicIntegBase,
-    BuildIntegNodeBase,
     BuildIntegGoBase,
-    BuildIntegProvidedBase,
     BuildIntegPythonBase,
-    BuildIntegJavaBase,
-    BuildIntegEsbuildBase,
 )
 
 
@@ -330,264 +324,6 @@ class TestBuildCommand_RubyFunctionsWithGemfileInTheRoot(BuildIntegRubyBase):
         # update template path with new location
         self.template_path = str(Path(self.working_dir).joinpath("template.yaml"))
 
-
-class TestBuildCommand_Java(BuildIntegJavaBase):
-    @parameterized.expand(
-        [
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-        ]
-    )
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    def test_building_java_in_container(
-        self, runtime, runtime_version, code_path, expected_files, expected_dependencies
-    ):
-        if not runtime_supported_by_docker(runtime):
-            self.skipTest(RUNTIME_NOT_SUPPORTED_BY_DOCKER_MSG)
-        self._test_with_building_java(
-            runtime,
-            os.path.join(code_path, runtime_version),
-            expected_files,
-            expected_dependencies,
-            "use_container",
-            self.test_data_path,
-        )
-
-    @parameterized.expand(
-        [
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java8.al2",
-                "8",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java11",
-                "11",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java17",
-                "17",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_GRADLE_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_GRADLEW_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_GRADLE_KOTLIN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                BuildIntegJavaBase.EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java21",
-                "21",
-                BuildIntegJavaBase.USING_MAVEN_PATH,
-                BuildIntegJavaBase.EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                BuildIntegJavaBase.EXPECTED_MAVEN_DEPENDENCIES,
-            ),
-        ]
-    )
-    def test_building_java_in_process(self, runtime, runtime_version, code_path, expected_files, expected_dependencies):
-        self._test_with_building_java(
-            runtime,
-            os.path.join(code_path, runtime_version),
-            expected_files,
-            expected_dependencies,
-            False,
-            self.test_data_path,
-        )
 
 
 class TestBuildCommand_Go_Modules(BuildIntegGoBase):
@@ -1048,92 +784,6 @@ class TestBuildCommand_LayerBuilds(BuildIntegBase):
 
     def _get_python_version(self):
         return "python{}.{}".format(sys.version_info.major, sys.version_info.minor)
-
-
-@parameterized_class(
-    ("template", "is_nested_parent"),
-    [
-        (os.path.join("nested-parent", "template-parent.yaml"), "is_nested_parent"),
-        ("template.yaml", False),
-    ],
-)
-class TestBuildCommand_ProvidedFunctions(BuildIntegProvidedBase):
-    # Test Suite for runtime: provided and where selection of the build workflow is implicitly makefile builder
-    # if the makefile is present.
-    @parameterized.expand(
-        [
-            ("provided", False, None),
-            ("provided", "use_container", "Makefile-container"),
-            ("provided.al2", False, None),
-            ("provided.al2", "use_container", "Makefile-container"),
-            ("provided.al2023", False, None),
-            ("provided.al2023", "use_container", "Makefile-container"),
-        ]
-    )
-    def test_building_Makefile(self, runtime, use_container, manifest):
-        if use_container:
-            if SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD:
-                self.skipTest(SKIP_DOCKER_MESSAGE)
-            if not runtime_supported_by_docker(runtime):
-                self.skipTest(RUNTIME_NOT_SUPPORTED_BY_DOCKER_MSG)
-        self._test_with_Makefile(runtime, use_container, manifest)
-
-
-@parameterized_class(
-    ("template", "is_nested_parent"),
-    [
-        (os.path.join("nested-parent", "template-parent.yaml"), "is_nested_parent"),
-        ("template.yaml", False),
-    ],
-)
-class TestBuildCommand_ProvidedFunctions_With_Specified_Architecture(BuildIntegProvidedBase):
-    # Test Suite for runtime: provided and where selection of the build workflow is implicitly makefile builder
-    # if the makefile is present.
-    @parameterized.expand(
-        [
-            ("provided", False, None, "x86_64"),
-            ("provided", "use_container", "Makefile-container", "x86_64"),
-            ("provided.al2", False, None, "x86_64"),
-            ("provided.al2", "use_container", "Makefile-container", "x86_64"),
-            ("provided.al2023", False, None, "x86_64"),
-            ("provided.al2023", "use_container", "Makefile-container", "x86_64"),
-        ]
-    )
-    def test_building_Makefile(self, runtime, use_container, manifest, architecture):
-        if use_container and not runtime_supported_by_docker(runtime):
-            self.skipTest(RUNTIME_NOT_SUPPORTED_BY_DOCKER_MSG)
-        self._test_with_Makefile(runtime, use_container, manifest, architecture)
-
-
-@parameterized_class(
-    ("template", "code_uri", "is_nested_parent"),
-    [
-        ("custom_build_with_custom_root_project_path.yaml", "empty_src_code", False),
-        ("custom_build_with_custom_make_file_path.yaml", "provided_src_code_without_makefile", False),
-        ("custom_build_with_custom_working_dir.yaml", "custom_working_dir_src_code", False),
-        ("custom_build_with_custom_root_project_path_and_custom_makefile_path.yaml", "empty_src_code", False),
-        (
-            "custom_build_with_custom_root_project_path_custom_makefile_path_and_custom_working_dir.yaml",
-            "empty_src_code",
-            False,
-        ),
-    ],
-)
-class TestBuildCommand_ProvidedFunctionsWithCustomMetadata(BuildIntegProvidedBase):
-    # Test Suite for runtime: provided and where selection of the build workflow is implicitly makefile builder
-    # if the makefile is present.
-    @parameterized.expand(
-        [
-            ("provided", False, None),
-            ("provided.al2", False, None),
-            ("provided.al2023", False, None),
-        ]
-    )
-    def test_building_Makefile(self, runtime, use_container, manifest):
-        if use_container and not runtime_supported_by_docker(runtime):
-            self.skipTest(RUNTIME_NOT_SUPPORTED_BY_DOCKER_MSG)
-        self._test_with_Makefile(runtime, use_container, manifest)
-
 
 @skipIf(
     ((IS_WINDOWS and RUNNING_ON_CI) and not CI_OVERRIDE),
